Sum of distance Minimal Matching cost
- h1 (Distance to Nearest Box): This heuristic calculates the Minkowski distance
from the player's current position to the nearest box that is not yet in a storage location.

- h2 (Minimum Cost Perfect Matching for Boxes to Goals): This heuristic computes the
Minkowski distances of all boxes to all storage locations and finds the minimum cost perfect
matching on a complete bipartite graph.

While individual heuristics h1 and h2 might be admissible on their own,
adding admissible heuristics does not guarantee that the resulting sum will also be admissible.